Refactored getSession into SessionController, cleaned exports and types
